backend/backend_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==36.0.2
distro==1.9.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.77.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
